Added Collection page component, CollectionOverview component, Directory and shop reducers.
Added Directory and Shop reducers to handle the state and data for both.
Added the CollectionOverview component for rendering each collection items available in its own page.
Added the redux-persist module to persist data for items added to the cart.
Made changes in the cart reducer for simplicity and persisted data for each user data.
Modified root-reducers and App.js to implement the use of redux-persist and also to configure the data
to persist.